Split up core create
