+++ test/core/core.spec.ts
@@ -559,74 +559,4 @@ contract("Core", (accounts) => {
       });
     });
   });
-
-  describe("#create", async () => {
-    let factoryAddress: Address;
-    let components: Address[];
-    const units: BigNumber[] = [ONE];
-    const naturalUnit: BigNumber = ONE;
-    const name = "New Set";
-    const symbol = "SET";
-
-    beforeEach(async () => {
-      await deployCoreAndInitializeDependencies();
-      mockToken = await coreWrapper.deployTokenAsync(ownerAccount);
-
-      factoryAddress = setTokenFactory.address;
-      components = [mockToken.address];
-    });
-
-    async function subject(): Promise<string> {
-      return core.create.sendTransactionAsync(
-        factoryAddress,
-        components,
-        units,
-        naturalUnit,
-        name,
-        symbol,
-        { from: ownerAccount },
-      );
-    }
-
-    it("creates a new SetToken and tracks it", async () => {
-      const txHash = await subject();
-
-      const logs = await getFormattedLogsFromTxHash(txHash);
-      const newSetTokenAddress = extractNewSetTokenAddressFromLogs(logs);
-
-      const isSetTokenValid = await core.validSets.callAsync(newSetTokenAddress);
-      expect(isSetTokenValid).to.be.true;
-    });
-
-    it("emits a SetTokenCreated event", async () => {
-      const txHash = await subject();
-      const logs = await getFormattedLogsFromTxHash(txHash);
-      const newSetTokenAddress = extractNewSetTokenAddressFromLogs(logs);
-
-      const expectedLogs: Log[] = [
-        SetTokenCreated(
-          core.address,
-          newSetTokenAddress,
-          factoryAddress,
-          components,
-          units,
-          naturalUnit,
-          name,
-          symbol,
-        ),
-      ];
-
-      await assertLogEquivalence(expectedLogs, logs);
-    });
-
-    describe("when the factory is not valid", async () => {
-      beforeEach(async () => {
-        factoryAddress = NULL_ADDRESS;
-      });
-
-      it("should revert", async () => {
-        await expectRevertError(subject());
-      });
-    });
-  });
 });
